refactor(daemon): extract shell/gh exec helpers into shell/commands.ts
Strangler-fig slice 2, cluster 3. Moves the buffered-subprocess and
POSIX login-shell command helpers — execFileBuffered, quotePosixShellArg,
buildGhShellCommand, buildCommandShellCommand, buildLoginShellCommand,
execGhBuffered, execCommandViaLoginShell — out of server.ts into a new
shell/commands.ts module. ensureGhReady moves with them (it is its only
consumer of execGhBuffered).

Byte-identical move. All eight were module-private in server.ts (nothing
external imported them), so no re-export is needed; server.ts imports
execCommandViaLoginShell back for its route deps object and the inline
plugin-share route. The execFile import stays — other server.ts code
still uses it directly.

Two adjacent findings, deliberately NOT changed here (follow-ups):
- ensureGhReady has no callers repo-wide (cli.ts does its own inline
  gh-readiness checks) — dead code, flagged for a removal PR.
- cli.ts keeps its own independent copies of these exec helpers; deduping
  both onto shell/commands.ts is a separate refactor.
Also dropped a stale orphaned JSDoc block (code/message/init -> ApiError)
that sat above execFileBuffered but described neither it nor its
neighbour; removing it during the move avoids re-attaching it to
readProjectPluginManifest.

server.ts 6004 -> 5926 lines. Public/runtime behavior unchanged.

Validation: tsc -p apps/daemon/tsconfig.json --noEmit clean; pnpm guard
clean; chat-run-sse-shapes characterization 4/4; no daemon test
references the moved symbols.
